remove blanks from receivers_email
The recipients of the email are extracted from a variable using the
split() function on a blank space (' '). If there are additional blanks,
either before, after or in the middle of the receivers_email variable the
sendmail function will cause an exception. This patch makes sure that blanks
won't cause the exception.