update (npm): add v0.2 with package_manager param
npm task's behavior defaults to using the npm command. Users
would like to use the same task with yarn or some other package manager
as their node env based package manager if npm is not their choice of
pkg manager. To enable this, this commit allows the user to specify the
COMMAND param (which default to npm) where the user can specify some
other package manager available in the nodejs environment.