Smaller default image for github-set-status
The default image for the github-set-status is more
than 800MB. The python code actually run in this task
is small and used only the stdlib. A smaller image
would works just as well.

The new default is python-alpine (45MB), which is the
smallest python image possible. It is now configurable.