Remove securityContext from buildpacks task
In buildpacks task, we were running prepare step as privileged because
the task wasn't able to run properly on RBAC enforced clusters but now
it seems that the barrier has been lowered by those clusters and the
task now works properly without it.

Bumping the version so that the task version 0.3 doesn't breaks for
existing users.

Signed-off-by: vinamra28 <vinjain@redhat.com>