Update e2e test script to Pipelines v0.32.0
Prior to this commit we were using Pipelines 0.28 for running e2e tests
in the Catalog.

This commit updates the pipelines version used during testing to 0.32.0,
which was released today.

After making this change one of the steps in an Orka Task is now failing
due to a naming conflict. This commit also renames the step to work
around the issue.