Versioning Resources
Today, we don't always bump versions of resources when updating them.
The guideline has been to bump versions only when behavior changes,
but it's hard to figure out when the behavior has changed (a change
that could be trivial to one user could be meaningful to another).

Not bumping resource versions when changing them causes issues where
the resource definition becomes dependent on the time when it was applied
 by the user - which causes unexpected failures as described in
https://github.com/tektoncd/catalog/issues/784.

This issue also came up as an issue where users cannot depend on the
Step indices because they can change:
https://github.com/tektoncd/community/pull/572#pullrequestreview-836578705.

In TEP-0003, we already proposed that a policy for versioning of resources:
https://github.com/tektoncd/community/blob/main/teps/0003-tekton-catalog-organization.md#versioning-resources

In Catalog Working Group on 01/13/2022, we revisited that policy and:
- agreed to follow the versioning policy
- make it easier to bump resources (see https://github.com/tektoncd/plumbing/pull/994)

This change documents the versioning policy in the contributions guide.

Fixes https://github.com/tektoncd/catalog/issues/784