gcs-upload v0.2 do not delete missing files by default
The gcs-upload task v0.1, when dealing with folders, uses the "-d"
option by default, which deletes remote files by default. This is
a relatively dangerous choice of default, and it's safer to make
that as an opt-in, as it may result in deletion of a lot of remote
files by accident.

V0.2 has a new parameter "deleteExtraFiles" which corresponds to
gsutil "-d" flag. deleteExtraFiles defaults to false, making the
default behaviour of the task different from v0.1.

V0.2 replaces parameters in the script with environment variables
for security as recommended in the catalog best practices.

The cloud-sdk image version has been updated to 390.0.0-slim which
is newer, ~60% smaller than the current image and includes gsutil.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>