Improve gcs-upload performace with busy targets
gcs-upload uses rsync to upload folders. When the target contains
many thousands of files, it can take a very long time.

Change the default implementation to use "cp -r" which only copies
new files. Recursive copy supports not replacing existing files,
so adding an option to the task for that.

The option to delete extra remote files is still there, and still
false by default. Delete extra remote files is not compatible with
"not replace remote files" so adding a note about that.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>