Use latest release for e2e tests
This commit updates the catalog e2e tests to use the latest Pipelines release.
Because this release removes support for v1alpha1 Task, TaskRun, Pipeline, and
PipelineRun, and the previous release removes support for Conditions, this
commit also updates tests to use up-to-date versions of Tekton CRDs.
It does not update any catalog Tasks or Pipelines.