Set HOME to /tekton/home for nonroot tasks
In an effort to reduce permissions for certain tasks in #1034, the HOME
directory was either set to `/home/nonroot` or was unchanged. This
caused the tasks to fail when being run under limited permissions as
they could not create `/home/nonroot` or write to `/` in some cases.

This commit follows suit with #860 where HOME is set to `/tekton/home`
so that the task can perform actions in it.