Add DELETE option to github-add-comment
This will be used as part of https://github.com/tektoncd/plumbing/issues/483 in Tekton's own CI setup, and I think it could be generally useful for replicating behavior such as Prow's job failure comments. Those comments are deleted once the job completes next, with a new comment created for subsequent failures, so that the latest result is always towards the end of the comment list.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>