Add namespace support in blue-green-deployment
As of now blue-green-deployment task assumes the namespace to be the
current namespace in which pipeline is running but everytime this might
not be the case. Added an extra parameter, `NAMESPACE` which user needs
to supply as the parameter in which the existing deployment and service
are present.

Also added an optional parameter `IMAGE` in order to improve the
flexibility for the user

Signed-off-by: vinamra28 <jvinamra776@gmail.com>