Switch container images from gcr.io to ghcr.io
Replace all references to images on gcr.io with the corresponding
copy on ghcr.io. The target images are identical to the source ones,
so there is no functional change in tasks or stepaction.

This PR changes old versions of tasks too and no new version is
created, which should make it easier for people to start pulling
from ghcr.io instead of gcr.io.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>