Switch to GHA for e2e tests
Adapt the setup-kind script from the pipeline repo.
Created a GHA workflow that can use the above script to
setup a k8s cluster and run the e2e script on top as is
done today on a prow cluster.

The GHA workflow also includes a matrix which will test
modified resources against the four supported LTS versions.

The list of required jobs needs to be updated in the
plumbing repository to match the new job name.

This PR also deprecates a number of tasks that should have
been deprecated before, as newer versions took precedence.

Skip the "kind" test (for now) because of nested
virtualization issues.

Unit tests are not needed, there's no unit test to run.
Add a lint job that checks go and yaml files and licenses.

The catlin lint job is implemented as a Tekton job and needs
to be migrated to GHA or setup on OCI.

The kind cluster includes a pass-through cache for docker.io
to help avoid rate limiting by pulling an image only once even
if it is requested by all nodes in the cluster.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>